New feature: Remember the cursor position within the argument
Motivational example:

sss -> grep '^'

You really want the cursor to not appear at the end of the string.
An easy and easily understandable way of doing this is remembering
the cursor position when defining snippets.

Other snippet dsls (like yas snippet) use a placeholder syntax.
The advantage of this approach is that you can *multiple* insertion
points along with default values.
There's a trade-off here between "do what I mean" and "power".

I don't really want to implement a snippet dsl in zsh...